% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_create.R
\name{data_create}
\alias{data_create}
\title{Creates a data folder}
\usage{
data_create(raw = FALSE, cleaned = FALSE)
}
\arguments{
\item{raw}{if you would like a "data-raw" folder, specify TRUE}

\item{cleaned}{if you woud like a "data-cleaned" folder, specify TRUE}
}
\value{

}
\description{
Creates a data folder
}
