% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_create.R
\name{output_create}
\alias{output_create}
\title{Creates an output file}
\usage{
output_create(fig = FALSE, report = FALSE)
}
\arguments{
\item{fig}{if you would like a figure folder, specify TRUE}

\item{report}{if you would like a report folder, specify TRUE}
}
\value{

}
\description{
Creates an output file
}
