% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_create.R
\name{src_create}
\alias{src_create}
\title{Creates a source file}
\usage{
src_create(Rsub = FALSE)
}
\arguments{
\item{Rsub}{if you want a source file for R, specify = TRUE}
}
\value{

}
\description{
Creates a source file
}
